refactor(ec2fleet): split deployer.go into focused domain files
Mirrors the gamelift deployer split pattern. No public API or behavior changes.

- deployer.go: struct, constructor, constants, resourceTags, orchestration (CreateFleet, Destroy, GetFleetStatus, sessions)
- deployer_fleet.go: fleet lifecycle (createFleetInput, runtimeConfiguration, waitForFleetActive, pollFleetActiveStatus, fleetActivePollResult, deleteFleetResource, waitForFleetDeletion, deleteBuildResource)
- deployer_iam.go: ensureIAMRole, deleteIAMRole
- deployer_s3.go: resolveBucket, deleteS3Object
