+++ internal/ec2fleet/deployer_iam.go
@@ -1,0 +1,106 @@
+package ec2fleet
+
+import (
+	"context"
+	"fmt"
+	"time"
+
+	"github.com/aws/aws-sdk-go-v2/aws"
+	"github.com/aws/aws-sdk-go-v2/service/iam"
+	"github.com/devrecon/ludus/internal/awsutil"
+	"github.com/devrecon/ludus/internal/tags"
+)
+
+// ensureIAMRole creates the GameLift EC2 fleet IAM role if it doesn't exist.
+func (d *Deployer) ensureIAMRole(ctx context.Context) (string, error) {
+	getOut, err := d.iamClient.GetRole(ctx, &iam.GetRoleInput{
+		RoleName: aws.String(iamRoleName),
+	})
+	if err == nil {
+		return aws.ToString(getOut.Role.Arn), nil
+	}
+
+	assumeRolePolicy := `{
+  "Version": "2012-10-17",
+  "Statement": [{
+    "Effect": "Allow",
+    "Principal": {"Service": "gamelift.amazonaws.com"},
+    "Action": "sts:AssumeRole"
+  }]
+}`
+
+	createOut, err := d.iamClient.CreateRole(ctx, &iam.CreateRoleInput{
+		RoleName:                 aws.String(iamRoleName),
+		AssumeRolePolicyDocument: aws.String(assumeRolePolicy),
+		Description:              aws.String("IAM role for Ludus GameLift EC2 fleet"),
+		Tags:                     tags.ToIAMTags(d.resourceTags()),
+	})
+	if err != nil {
+		return "", fmt.Errorf("creating IAM role: %w", err)
+	}
+
+	_, err = d.iamClient.AttachRolePolicy(ctx, &iam.AttachRolePolicyInput{
+		RoleName:  aws.String(iamRoleName),
+		PolicyArn: aws.String(iamPolicyARN),
+	})
+	if err != nil {
+		return "", fmt.Errorf("attaching policy to role: %w", err)
+	}
+
+	// EC2 managed builds require S3 read access for GameLift to download
+	// the build archive. The GameLiftContainerFleetPolicy only covers
+	// container fleets, so we add an inline policy for S3.
+	s3Policy := `{
+  "Version": "2012-10-17",
+  "Statement": [{
+    "Effect": "Allow",
+    "Action": ["s3:GetObject", "s3:GetObjectVersion"],
+    "Resource": "arn:aws:s3:::ludus-builds-*/*"
+  }]
+}`
+	_, err = d.iamClient.PutRolePolicy(ctx, &iam.PutRolePolicyInput{
+		RoleName:       aws.String(iamRoleName),
+		PolicyName:     aws.String("LudusS3BuildAccess"),
+		PolicyDocument: aws.String(s3Policy),
+	})
+	if err != nil {
+		return "", fmt.Errorf("adding S3 access policy: %w", err)
+	}
+
+	// Wait for IAM propagation
+	time.Sleep(10 * time.Second)
+
+	return aws.ToString(createOut.Role.Arn), nil
+}
+
+func (d *Deployer) deleteIAMRole(ctx context.Context) error {
+	fmt.Println("Deleting IAM role...")
+
+	_, err := d.iamClient.DetachRolePolicy(ctx, &iam.DetachRolePolicyInput{
+		RoleName:  aws.String(iamRoleName),
+		PolicyArn: aws.String(iamPolicyARN),
+	})
+	if err != nil && !awsutil.IsNotFound(err) {
+		return fmt.Errorf("detaching policy from role: %w", err)
+	}
+
+	// Remove S3 access inline policy
+	_, _ = d.iamClient.DeleteRolePolicy(ctx, &iam.DeleteRolePolicyInput{
+		RoleName:   aws.String(iamRoleName),
+		PolicyName: aws.String("LudusS3BuildAccess"),
+	})
+
+	_, err = d.iamClient.DeleteRole(ctx, &iam.DeleteRoleInput{
+		RoleName: aws.String(iamRoleName),
+	})
+	if err != nil {
+		if awsutil.IsNotFound(err) {
+			fmt.Println("IAM role not found, skipping.")
+			return nil
+		}
+		return fmt.Errorf("deleting IAM role: %w", err)
+	}
+
+	fmt.Println("IAM role deleted.")
+	return nil
+}
